# openapi.yaml
openapi: 3.0.0
info:
  title: Main Collection
  version: 1.0.0
  description: Contains the Postman docs for all of our api/endpoionts.
servers:
  - url: http://localhost:3000
paths:
...
  /api/threads/:
    post:
      summary: Create Thread
      responses:
        default:
          description: Create Thread with Tags
      tags:
        - Threads
      security:
        - BearerAuth: []
      description: "## Get Threads\n\n**GET** `/api/threads`\n\nRetrieves a paginated list of threads with optional filters.\n\n### Query Parameters\n\n- `title` (optional): Filter threads by title (contains)\n- `author` (optional): Filter by author username\n- `team` (optional): Filter by team name\n- `tags` (optional): Filter by tag names (can be comma-separated or multiple query params)\n- `page` (optional, default: 1): Page number for pagination\n- `pageSize` (optional, default: 10): Number of items per page\n\n### Example Request\n\n```\nGET /api/threads?page=1&pageSize=10&team=Lakers&tags=game-thread,playoffs\n```\n\n### Success Response\n\n**Code:** `200 OK`\n\n```json\n[\n\t{\n\t\t\"id\": 1,\n\t\t\"title\": \"Lakers vs Warriors Discussion\",\n\t\t\"createdById\": 5,\n\t\t\"teamId\": 3,\n\t\t\"mainPostId\": 42,\n\t\t\"isVisible\": true,\n\t\t\"isClosed\": false,\n\t\t\"matchId\": null,\n\t\t\"opensAt\": null,\n\t\t\"closesAt\": null,\n\t\t\"createdAt\": \"2026-03-06T10:00:00.000Z\",\n\t\t\"updatedAt\": \"2026-03-06T10:00:00.000Z\",\n\t\t\"tags\": [\n\t\t\t{\n\t\t\t\t\"tag\": {\n\t\t\t\t\t\"id\": 1,\n\t\t\t\t\t\"name\": \"game-thread\"\n\t\t\t\t}\n\t\t\t}\n\t\t]\n\t}\n]\n```\n\n### Error Responses\n\n**Code:** `400 Bad Request`\n\n```json\n{\n\t\"error\": \"Page and pageSize must be positive integers.\"\n}\n```\n\n**Code:** `404 Not Found`\n\n```json\n{\n\t\"error\": \"Author not found.\"\n}\n```\n\n```json\n{\n\t\"error\": \"Team not found.\"\n}\n```\n\n**Code:** `500 Internal Server Error`\n\n```json\n{\n\t\"error\": \"Error fetching threads.\"\n}\n```\n\n### Notes\n\n- Only returns visible threads\n- For threads associated with matches, only returns threads that are currently open (between opensAt and closesAt)\n- Tags can be specified multiple ways:\n    - `?tags=tag1,tag2` (comma-separated)\n    - `?tags=tag1&tags=tag2` (multiple params)"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                mainPostId:
                  type: string
            example:
              title: Sample Title
              mainPostId: ''
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
security:
  - BearerAuth: []